Fix hero img The Hague
In styles.css and hague.css I found 3 different links for hero background images. I changed one of them to the image with better resolution and deleted two more.
Also I have added text-shadow for paragraph and heading in order to make them readable and changed a bit height limitation for the image.